pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "maxbridge"
version = "0.1.0a0"
description = "Local-first interoperability client and Telegram forum bridge toolkit for MAX."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
  { name = "MAXBRIDGE contributors" },
]
keywords = [
  "asyncio",
  "bridge",
  "local-first",
  "max",
  "telegram",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Communications :: Chat",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "aiosqlite>=0.20",
  "httpx>=0.27",
  "orjson>=3.10",
  "pydantic>=2.7",
  "PyYAML>=6.0.1",
  "python-telegram-bot>=22.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "mkdocs-material>=9.5",
  "pre-commit>=3.7",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]

[project.scripts]
maxbridge = "maxbridge.cli.main:run"

[tool.hatch.build.targets.wheel]
packages = ["maxbridge"]
include = [
  "maxbridge/storage/migrations/*.sql",
  "maxbridge/py.typed",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
  "ASYNC",
  "RUF",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
plugins = []

[tool.coverage.run]
branch = true
source = ["maxbridge"]
